test: maximize FuSa coverage; fix wire frame-length integer overflow
Add comprehensive requirements, tests, and fuzzing across the project,
and fix a denial-of-service bug found by fuzzing the wire decoders.

wire (NEW: 0% -> 100% coverage):
- Add REQ-WIRE-001..009 covering header validation, Command/Response/
  Status round-trips, truncated-body rejection, control frames, and
  decoder fuzz-robustness at the transport trust boundary.
- Fix integer overflow in DecodeCommand/DecodeResponse/DecodeStatus:
  a crafted bodyLen near math.MaxUint32 wrapped 'uint32(HeaderLen)+bodyLen'
  to a small value, defeating the truncation guard and panicking with an
  out-of-bounds slice. Any UDP/TLS transport could be crashed by one
  malformed frame. Now computed in uint64. Caught by FuzzDecode*.

Coverage gains in core and supporting packages:
- root (go-RCP): 78.3% -> 100% (adapt Send/Call/Subscribe error and
  back-pressure paths, CloseWithDrain drain/expiry, Loan/NewLoan, Zone.String).
- shmem: 65.0% -> 95.7% (Registry interface, accessors, SetHealthy).
- loan: 67.7% -> 93.5% (delegating accessors, closed/large/pool-reuse paths).
- mdns: 71.1% -> 78.9% (dns.go parsing primitives + FuzzParseRRs).
- admin: 78.8% -> 89.8% (parseZone, RecordSend telemetry, events, metrics).
- restbridge: 73.0% -> 82.0% (JSON/controller error paths, name parseZone).
- grpcbridge: 76.4% -> 79.2% (client Zone, Subscribe-after-close).

All gates green: gofusa check, golangci-lint v2, go vet, go test -race ./...

+++ wire/wire.go
@@ -1,6 +1,16 @@
 // Package wire defines the shared RCP binary frame format used by both UDP and TLS transports.
 package wire
 
+//fusa:req REQ-WIRE-001
+//fusa:req REQ-WIRE-002
+//fusa:req REQ-WIRE-003
+//fusa:req REQ-WIRE-004
+//fusa:req REQ-WIRE-005
+//fusa:req REQ-WIRE-006
+//fusa:req REQ-WIRE-007
+//fusa:req REQ-WIRE-008
+//fusa:req REQ-WIRE-009
+
 import (
 	"encoding/binary"
 	"errors"
@@ -65,7 +75,9 @@ func DecodeCommand(b []byte) (*rcp.Command, error) {
 		return nil, err
 	}
 	bodyLen := binary.BigEndian.Uint32(b[12:16])
-	if uint32(len(b)) < uint32(HeaderLen)+bodyLen {
+	// uint64 arithmetic avoids the wraparound that would otherwise let a
+	// crafted bodyLen near math.MaxUint32 pass this guard and panic the slice.
+	if uint64(len(b)) < uint64(HeaderLen)+uint64(bodyLen) {
 		return nil, ErrShortFrame
 	}
 	cmd := &rcp.Command{
@@ -102,7 +114,7 @@ func DecodeResponse(b []byte) (*rcp.Response, error) {
 		return nil, err
 	}
 	bodyLen := binary.BigEndian.Uint32(b[12:16])
-	if uint32(len(b)) < uint32(HeaderLen)+bodyLen {
+	if uint64(len(b)) < uint64(HeaderLen)+uint64(bodyLen) {
 		return nil, ErrShortFrame
 	}
 	resp := &rcp.Response{
@@ -140,7 +152,7 @@ func DecodeStatus(b []byte) (*rcp.Status, error) {
 		return nil, err
 	}
 	bodyLen := binary.BigEndian.Uint32(b[12:16])
-	if uint32(len(b)) < uint32(HeaderLen)+bodyLen {
+	if uint64(len(b)) < uint64(HeaderLen)+uint64(bodyLen) {
 		return nil, ErrShortFrame
 	}
 	st := &rcp.Status{
